Fix MissingBodyAnnotationTest
The test in MissingBodyAnnotationTest has a request with a POJO as a body and no content type.

When converting this to a Google Request prior to this PR we were simply calling toString() on
the POJO, which resulted in an invalid body in the Google request.

This PR attempts to convert the body to a usable form instead.

I think this is ok, but I am only about 90% sure this is the correct fix, as it differs from the
AWS solution for the same issue, so I worry that toGoogleRequest is used elsewhere in the stack.

In AWS it's only used in testing