GCP auth failure logging and testing
AOP around advice is applied to managed instances of GoogleCredentials
to warn when there are failures that will be silently handled by the
GCP libraries as an infinite series of retry requests.

A thorough set of tests are added to validate the algorithm in
GoogleCredentialsFactory for locating GCP credentials and ensure that
the new AOP interceptor works with the different subclass variations of
GoogleCredentials.