Add 'nack-on-shutdown' configuration option
A configuration property is added to allow the application shutdown
behavior to be modified such that pending in-memory PubSub messages are
eagerly nacked instead of processing to completion.

Tests are added to ensure the prior default behavior is retained and
that the new configuration property works as intended.

Documentation is updated to explain the default shutdown behavior as
well as the new configuration option.