Add support for reactive PubSub consumers
Support for reactive consumers is added. This gives the ability to bind
the received PubSub message as a reactive type to the subscriber
method's body parameter, and to return a reactive type to which the
framework will subscribe.

Message automatic and manual ack/nack are handled accordingly.

Docs are added for the added reactive support.

This resolves issue #945