Refactor SubscriberShutdownSpec
`SubscriberShutdownSpec` has been refactored to use a test replacement
for `DefaultSubscriberFactory` that enables comparing the number of
messages received from PubSub to the number of messages processed by the
listener in the test. This should allow tolerance for latency between
the test environment and the PubSub emulator Docker container which
could cause the number of messages processed not being an exact match
to the number of messages published due to the way the GCP library
shuts down the GRPC stream and is acceptable behavior, as we are just
trying to verify the processing of all messages that have been received
and buffered in memory at shutdown time.