Stop trying to guess a port for Jetty in http-tests
This changes how we generate a port for Jetty in the http-tests.
Previously we tried to guess one then tried to start jetty on it in a retry loop.
After this change, we just let Jetty pick a port by assigning it 0.

This wasn't an issue in this module, but it has proved to be a source of flakiness in Azure, so I'm doing this for consistency