Add HttpClient based implementation of HttpTransportFactory
An implementation of the GCP SDK API `HttpTransportFactory` is added that internally uses the default Micronaut
`HttpClient`. This API is used by the GCP `GoogleCredentials` for refreshing OAuth tokens for use in API calls. This
implementation adds enhanced logging of token refresh requests that is useful in debugging misconfigured
credentials. Using the default `HttpClient` also allows the potential for setting timeouts and other such properties
via typical Micronaut configuration.

The implementation uses the blocking version of the `HttpClient` API. This is safe to use in this case as the GCP
library is already executing these tasks using its own managed threads, so these requests should not block the main
Micronaut event loop.

The implementation is only activated if `HttpClient` is found to be on the classpath. If it is not, the default
implementation provided by the SDK will be used instead.